change grammar so that expr_init does not cause amgibuity
The expr_init was ambiguous in places which were followed by braces,
e.g. ifs or loops. The braces following those statements were parsed as
expr_init instead and caused errors. We change the grammar so that
expr_init is not directly allowed in those context (i.e. must be
parenthesised).