replace ast::Decl_Overloaded_Function with ast::Overload_Group
Remove ast::Decl_Overloaded_Function and in its place create
ast::Overload_Group. The new struct is not an AST node. Instead, it
serves the purpose of an owning grouping of overloads, so that
Decl_Functions may be preserved in the original AST.

Additionally, we take advantage of the mutable AST and add attributes to
the nodes to remove the need for lookup tables.